This was silly and I should feel silly

+++ src/main/java/com/me/tft_02/duel/util/player/DuelManager.java
@@ -154,6 +154,9 @@ public class DuelManager {
     public static void endDuel(Player player, boolean deleteArena) {
         PlayerData.removeDuelTarget(player);
         PlayerData.setOccupied(player, false);
-        ArenaManager.deleteArena(player);
+
+        if (deleteArena) {
+            ArenaManager.deleteArena(player);
+        }
     }
 }
